# frontend/api/api-spec.yaml
openapi: 3.1.0
info:
    title: Kindred API
    description: Kindred API built with Huma v2 and Fiber
    version: 1.0.0
paths:
...
    /v1/user:
        get:
            tags:
                - auth
            summary: Test authentication
            description: Test endpoint for authenticated users
            operationId: auth-test
            parameters:
                - name: Authorization
                  in: header
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TestOutputBody'
                default:
                    description: Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ErrorModel'
    /v1/user/:
        get:
            tags:
                - auth
            summary: Test authentication
            description: Test endpoint for authenticated users
            operationId: auth-test-slash
            parameters:
                - name: Authorization
                  in: header
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TestOutputBody'
                default:
                    description: Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ErrorModel'
...
    /welcome:
        get:
            tags:
                - general
            summary: Welcome endpoint
            description: Returns a welcome message
            operationId: get-welcome
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/WelcomeOutputBody'
                default:
                    description: Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ErrorModel'
components:
    schemas:
...
        ErrorDetail:
            type: object
            additionalProperties: false
            properties:
                location:
                    type: string
                    description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
                message:
                    type: string
                    description: Error message text
                value:
                    description: The value at the given location
        ErrorModel:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://example.com/schemas/ErrorModel.json
                    readOnly: true
                detail:
                    type: string
                    description: A human-readable explanation specific to this occurrence of the problem.
                    examples:
                        - Property foo is required but is missing.
                errors:
                    type: array
                    description: Optional list of individual error details
                    items:
                        $ref: '#/components/schemas/ErrorDetail'
                instance:
                    type: string
                    description: A URI reference that identifies the specific occurrence of the problem.
                    format: uri
                    examples:
                        - https://example.com/error-log/abc123
                status:
                    type: integer
                    description: HTTP status code
                    format: int64
                    examples:
                        - 400
                title:
                    type: string
                    description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
                    examples:
                        - Bad Request
                type:
                    type: string
                    description: A URI reference to human-readable documentation for the error.
                    format: uri
                    default: about:blank
                    examples:
                        - https://example.com/errors/example
...
        TestOutputBody:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://example.com/schemas/TestOutputBody.json
                    readOnly: true
                message:
                    type: string
                    examples:
                        - Authorized!
            required:
                - message
...
        WelcomeOutputBody:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://example.com/schemas/WelcomeOutputBody.json
                    readOnly: true
                message:
                    type: string
                    examples:
                        - Welcome to Kindred!
            required:
                - message
...